-> 4.21

+++ osa-4/blogilista/controllers/blogs.js
@@ -12,7 +12,7 @@ blogsRouter.get('/', async (request, response) => {
 })
 
 blogsRouter.get('/:id', async (request, response) => {
-  const blog = await Blog.find({_id: request.params.id })
+  const blog = await Blog.find({_id: request.params.id }).populate('user')
   logger.info(blog)
   response.json(blog)
 })
@@ -27,6 +27,7 @@ blogsRouter.post('/', async (request, response) => {
             error: 'token missing or invalid'
         })
     }
+
     const user = await User.findById(decodedToken.id)
     const finalBody = await { ...body, user: user.id }
     const blog = new Blog(finalBody)
@@ -35,6 +36,25 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete(`/:id`, async (req, res) => {
+    //Validoidaan oikeus poistaa kyseinen blogi
+    if (!req.token) {
+        return res.status(401).send({ error: 'missing token'})
+    }
+
+    //Etsitään blogin tekijän id
+    const blogToDel = await Blog.findById({_id: req.params.id })
+    const writerId = blogToDel.user ? blogToDel.user.toString() : null
+    logger.info('Writer id: ', writerId)
+
+    //Verrataan id:tä pyynnön mukana olevan tokenin id:seen
+    const reqId = jwt.verify(req.token, process.env.SECRET).id.toString()
+
+    logger.info('Request id: ', reqId)
+
+    if (reqId !== writerId) {
+        return res.status(401).send({ error: 'not allowed to delete'})
+    }
+
     const blog = await Blog.deleteOne({_id: req.params.id })
     logger.info('Deleted ', blog.deletedCount, ' blog')
     res.status(204).send('Blog deleted')
